refactor and add coverage
